+++ GHEtool/test/unit-tests/test_gfunction.py
@@ -2,10 +2,12 @@ import copy
 import time
 
 import numpy as np
+import pandas as pd
 import pygfunction as gt
+import pytest
 from pytest import raises
 
-from GHEtool import Borefield
+from GHEtool import Borefield, FOLDER
 from GHEtool.VariableClasses import FIFO, GFunction
 
 borehole_length_array = np.array([1, 5, 6])
@@ -457,3 +459,113 @@ def test_negative_values():
     g_func = gfunc.calculate(time, field, 1 / 5000 / 1000, interpolate=False)
     assert np.all(g_func > 0)
     assert np.isclose(np.min(g_func), 0.14299471464245733)
+
+
+def test_ann_boundaries():
+    gfunc = GFunction()
+    time_steps = np.arange(3600, 3600 * 24 * 365 * 100, 3600)
+    with pytest.warns():
+        borefield = gt.borefield.Borefield.rectangle_field(5, 1, 4, 6, 20, 0.5, 0.1)
+        gfunc.borefield = borefield
+        gfunc.calculate(time_steps, borefield, 1.1249999999999998e-06, use_neural_network=True,
+                        borefield_description={"type": 3, "N_1": 2, "N_2": 1,
+                                               "B_1": 4, "B_2": 6})
+    with pytest.warns():
+        borefield = gt.borefield.Borefield.rectangle_field(5, 1, 4, 6, 500, 0.5, 0.1)
+        gfunc.borefield = borefield
+        gfunc.calculate(time_steps, borefield, 1.1249999999999998e-06, use_neural_network=True,
+                        borefield_description={"type": 3, "N_1": 2, "N_2": 1,
+                                               "B_1": 4, "B_2": 6})
+    with pytest.warns():
+        borefield = gt.borefield.Borefield.rectangle_field(5, 1, 4, 6, 100, 8, 0.1)
+        gfunc.borefield = borefield
+        gfunc.calculate(time_steps, borefield, 1.1249999999999998e-06, use_neural_network=True,
+                        borefield_description={"type": 3, "N_1": 2, "N_2": 1,
+                                               "B_1": 4, "B_2": 6})
+    with pytest.warns():
+        borefield = gt.borefield.Borefield.rectangle_field(5, 1, 4, 6, 100, -2, 0.1)
+        gfunc.borefield = borefield
+        gfunc.calculate(time_steps, borefield, 1.1249999999999998e-06, use_neural_network=True,
+                        borefield_description={"type": 3, "N_1": 2, "N_2": 1,
+                                               "B_1": 4, "B_2": 6})
+    with pytest.warns():
+        borefield = gt.borefield.Borefield.rectangle_field(5, 1, 4, 6, 100, 0.5, 0.1)
+        gfunc.borefield = borefield
+        gfunc.calculate(time_steps, borefield, 1.1249999999999998e-06, use_neural_network=True,
+                        borefield_description={"type": 3, "N_1": 31, "N_2": 1,
+                                               "B_1": 4, "B_2": 6})
+    with pytest.warns():
+        borefield = gt.borefield.Borefield.rectangle_field(5, 1, 4, 6, 100, 0.5, 0.1)
+        gfunc.borefield = borefield
+        gfunc.calculate(time_steps, borefield, 1.1249999999999998e-06, use_neural_network=True,
+                        borefield_description={"type": 3, "N_1": 0, "N_2": 1,
+                                               "B_1": 4, "B_2": 6})
+    with pytest.warns():
+        borefield = gt.borefield.Borefield.rectangle_field(5, 1, 4, 6, 100, 0.5, 0.1)
+        gfunc.borefield = borefield
+        gfunc.calculate(time_steps, borefield, 1.1249999999999998e-06, use_neural_network=True,
+                        borefield_description={"type": 3, "N_1": 2, "N_2": 31,
+                                               "B_1": 4, "B_2": 6})
+    with pytest.warns():
+        borefield = gt.borefield.Borefield.rectangle_field(5, 1, 4, 6, 100, 0.5, 0.1)
+        gfunc.borefield = borefield
+        gfunc.calculate(time_steps, borefield, 1.1249999999999998e-06, use_neural_network=True,
+                        borefield_description={"type": 3, "N_1": 1, "N_2": 0, "B_1": 4, "B_2": 6})
+
+    with pytest.warns():
+        borefield = gt.borefield.Borefield.rectangle_field(5, 1, 4, 6, 100, 0.5, 0.1)
+        gfunc.borefield = borefield
+        gfunc.calculate(time_steps, borefield, 1.1249999999999998e-06, use_neural_network=True,
+                        borefield_description={"type": 3, "N_1": 1, "N_2": 2, "B_1": 1, "B_2": 6})
+    with pytest.warns():
+        borefield = gt.borefield.Borefield.rectangle_field(5, 1, 4, 6, 100, 0.5, 0.1)
+        gfunc.borefield = borefield
+        gfunc.calculate(time_steps, borefield, 1.1249999999999998e-06, use_neural_network=True,
+                        borefield_description={"type": 3, "N_1": 1, "N_2": 2, "B_1": 11, "B_2": 6})
+    with pytest.warns():
+        borefield = gt.borefield.Borefield.rectangle_field(5, 1, 4, 6, 100, 0.5, 0.1)
+        gfunc.borefield = borefield
+        gfunc.calculate(time_steps, borefield, 1.1249999999999998e-06, use_neural_network=True,
+                        borefield_description={"type": 3, "N_1": 1, "N_2": 2, "B_1": 2, "B_2": 1})
+    with pytest.warns():
+        borefield = gt.borefield.Borefield.rectangle_field(5, 1, 4, 6, 100, 0.5, 0.1)
+        gfunc.borefield = borefield
+        gfunc.calculate(time_steps, borefield, 1.1249999999999998e-06, use_neural_network=True,
+                        borefield_description={"type": 3, "N_1": 1, "N_2": 2, "B_1": 2, "B_2": 11})
+    with pytest.warns():
+        borefield = gt.borefield.Borefield.rectangle_field(5, 1, 4, 6, 100, 0.5, 0.01)
+        gfunc.borefield = borefield
+        gfunc.calculate(time_steps, borefield, 1.1249999999999998e-06, use_neural_network=True,
+                        borefield_description={"type": 3, "N_1": 1, "N_2": 2, "B_1": 2, "B_2": 2})
+    with pytest.warns():
+        borefield = gt.borefield.Borefield.rectangle_field(5, 1, 4, 6, 100, 0.5, 0.2)
+        gfunc.borefield = borefield
+        gfunc.calculate(time_steps, borefield, 1.1249999999999998e-06, use_neural_network=True,
+                        borefield_description={"type": 3, "N_1": 1, "N_2": 2, "B_1": 2, "B_2": 2})
+    with pytest.warns():
+        borefield = gt.borefield.Borefield.rectangle_field(5, 1, 4, 6, 100, 0.5, 0.1)
+        gfunc.borefield = borefield
+        gfunc.calculate(time_steps, borefield, 1e-8, use_neural_network=True,
+                        borefield_description={"type": 3, "N_1": 1, "N_2": 2, "B_1": 2, "B_2": 2})
+    with pytest.warns():
+        borefield = gt.borefield.Borefield.rectangle_field(5, 1, 4, 6, 100, 0.5, 0.1)
+        gfunc.borefield = borefield
+        gfunc.calculate(time_steps, borefield, 1e-5, use_neural_network=True,
+                        borefield_description={"type": 3, "N_1": 1, "N_2": 2, "B_1": 2, "B_2": 2})
+    with pytest.raises(ValueError):
+        borefield = gt.borefield.Borefield.rectangle_field(5, 1, 4, 6, 100, 0.5, 0.1)
+        gfunc.borefield = borefield
+        gfunc.calculate(time_steps, borefield, 1e-6, use_neural_network=True,
+                        borefield_description={"type": -1, "N_1": 1, "N_2": 2, "B_1": 2, "B_2": 2})
+    with pytest.raises(ValueError):
+        borefield = gt.borefield.Borefield.rectangle_field(5, 1, 4, 6, 100, 0.5, 0.1)
+        gfunc.borefield = borefield
+        gfunc.calculate(time_steps, borefield, 1e-6, use_neural_network=True,
+                        borefield_description={"type": 7, "N_1": 1, "N_2": 2, "B_1": 2, "B_2": 2})
+
+
+def test_no_borefield_description_ann():
+    gfunc = GFunction()
+    time_steps = np.arange(3600, 3600 * 24 * 365 * 100, 3600)
+    with pytest.raises(ValueError):
+        gfunc.calculate(time_steps, borefield, 1e-6, use_neural_network=True)